Separate File For Export
Separate class for the table in the N5 plugin.
